% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birds_eye.R
\name{birds_eye}
\alias{birds_eye}
\title{Function to make interactive plot}
\usage{
birds_eye(tidy_data, start, end, location)
}
\arguments{
\item{tidy_data}{tidy data from clean_food}

\item{start}{start date}

\item{end}{end date}

\item{location}{specified dining hall(s)}
}
\value{
A ggplot object
}
\description{
Function to make interactive plot
}
\examples{
\dontrun{
# Get data
data("raw_food")
df <- clean_food(raw_food)
tidy_data <- df[['tidy_data']]

# Call function
birds_eye(tidy_data, min(tidy_data$date), max(tidy_data$date), c("Tyler", "Comstock", "Cutter Ziskind", "Compass Cafe"))
}
}
