% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_food2.R
\name{clean_food2}
\alias{clean_food2}
\title{Function to clean raw FRN survey data}
\usage{
clean_food2(data)
}
\arguments{
\item{data}{corrected raw data from FRN survey}
}
\value{
A list with two data frames:
\describe{
\item{clean_data}{}
\item{tidy_data}{}
}
}
\description{
Function to clean raw FRN survey data
}
\examples{
\dontrun{
cooked_food <- clean_food2(data)
}
}
