% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_recovery.R
\name{plot_recovery}
\alias{plot_recovery}
\title{Function to plot recovery}
\usage{
plot_recovery(tidy_data, start, end, location, fill = FALSE)
}
\arguments{
\item{tidy_data}{tidy data from clean_food}

\item{start}{start date}

\item{end}{end date}

\item{location}{specified dining hall(s)}

\item{fill}{option to fill with food type if multiple locations are selected}
}
\value{
A ggplot object
}
\description{
Function to plot recovery
}
\examples{
\dontrun{
# Get data
data("raw_food")
cooked_food <- clean_food(raw_food)

# Call function
plot_recovery(cooked_food, '10-1-2024', '11-01-2024', "Tyler")
plot_recovery(cooked_food, '10-1-2024', '11-01-2024', c("Tyler", "Comstock"), fill = TRUE)
}
}
