% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{raw_food}
\alias{raw_food}
\title{raw_food: raw data from the Food Rescue Network Qualtrics survey
Contains information regarding dining halls, food type, and weight.}
\format{
A data frame with 238 rows and 248 variables:

Columns correspond to different Qualtrics questions.
Use \code{names(raw_food)} to learn more.
}
\usage{
raw_food
}
\description{
raw_food: raw data from the Food Rescue Network Qualtrics survey
Contains information regarding dining halls, food type, and weight.
}
\keyword{datasets}
