% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recovery_tables.R
\name{recovery_tables}
\alias{recovery_tables}
\title{Function to clean raw FRN survey data}
\usage{
recovery_tables(
  tidy_data,
  start,
  end,
  location,
  option = c("total", "type", "hall", "top")
)
}
\arguments{
\item{tidy_data}{tidy data from clean_food}

\item{start}{start date}

\item{end}{end date}

\item{location}{which dining halls are to be included}

\item{option}{total weight, grouped by food type, grouped by dining hall, and grouped by both food type and dining hall}
}
\value{
table
}
\description{
Function to clean raw FRN survey data
}
\examples{
\dontrun{
# Get data
data("raw_food")
df <- clean_food(raw_food)
tidy_data <- df[['tidy_data']]

# Call function
recovery_tables(tidy_data, '10-01-2024', '12-01-2024', "Tyler", "total")
recovery_tables(tidy_data, '10-01-2024', '12-01-2024', "Tyler", "type")
}
}
