% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weekly_stats.R
\name{weekly_stats}
\alias{weekly_stats}
\title{Function to clean raw FRN survey data}
\usage{
weekly_stats(clean_data, tidy_data, start, end, location)
}
\arguments{
\item{clean_data}{cleaned data from clean_food}

\item{tidy_data}{tidy data from clean_food}

\item{start}{start date}

\item{end}{end date}

\item{location}{specified dining hall(s)}
}
\value{
A list with three values: weekly summed weight (lb), # containers, # pick-ups
}
\description{
Function to clean raw FRN survey data
}
\examples{
\dontrun{
# Get data
data("raw_food")
df <- clean_food(raw_food)
tidy_data <- df[['tidy_data']]
clean_data <- df[['clean_data']]

# Call function
weekly_stats(clean_data, tidy_data, '10-1-2024', '11-1-2024', "Tyler")
}
}
